fix(repository): address Phase-1 code review issues in TableRelation
- Add @Transactional to @Modifying deleteByDataSourceIdAndSourceIn to
  prevent TransactionRequiredException at runtime
- Change isActive field type from Short to Integer in TableRelation
  model and repository parameter, consistent with TermGlossary and
  SqlExample entities
- Add default safelyFindRelationsInvolvingTables() to short-circuit
  empty-list calls and avoid invalid IN-clause SQL
- Add Javadoc on findRelationsInvolvingTables() warning callers about
  non-empty precondition

+++ src/main/java/com/example/mysqlbot/model/TableRelation.java
@@ -51,7 +51,7 @@ public class TableRelation {
 
     @Column(name = "is_active")
     @Builder.Default
-    private Short isActive = 1;
+    private Integer isActive = 1;
 
     @Column(name = "created_at", updatable = false)
     private LocalDateTime createdAt;
